+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Patch: Display shelving order on the item record. Refs UIIN-816.
 * Add <Pluggable> instance to the Instance action menu for plugin type `copyright-permissions-checker`.
 * Add callout after instance record is saved. Refs UIIN-1468.
+* Add callout after item or holdings record are saved. Refs UIIN-1485, UIIN-1486.
 
 ## [6.0.0](https://github.com/folio-org/ui-inventory/tree/v6.0.0) (2021-03-18)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.6...v6.0.0)